Fix e97bbeb9a76578c46fa190884c1601b2713cddb0 for source installs

+++ Library/Formula/python.rb
@@ -148,7 +148,7 @@ class Python < Formula
     site_packages.mkpath
 
     # Symlink the prefix site-packages into the cellar.
-    site_packages_cellar.unlink
+    site_packages_cellar.unlink if site_packages_cellar.exist?
     site_packages_cellar.parent.install_symlink site_packages
 
     # Write our sitecustomize.py
